+++ pyrope/core.py
@@ -2,12 +2,11 @@
 import abc
 import argparse
 import collections
-from functools import cache, cached_property
+from functools import cached_property
 import importlib
 import inspect
 import itertools
 import os
-from pathlib import Path
 import sys
 import unittest
 
@@ -17,6 +16,10 @@ import numpy
 from pyrope import frontends, tests
 from pyrope.config import process_total_score
 from pyrope.errors import IllPosedError
+from pyrope.messages import (
+    ChangeWidgetAttribute, CreateWidget, ExerciseAttribute, RenderTemplate,
+    Submit, WaitingForSubmission
+)
 
 
 float_types = (bool, int, float, numpy.bool_, numpy.int_, numpy.float_)
@@ -33,21 +36,12 @@ class Exercise(abc.ABC):
         self.weights = weights
         self.kwargs = kwargs
 
-    @property
-    def dir(self):
-        module = sys.modules[self.__module__]
-        try:
-            file = module.__file__
-        except AttributeError:
-            return Path.cwd()
-        return Path(file).parent
-
     def run(self, debug=False):
-        runner = ExerciseRunner(self)
+        runner = ExerciseRunner(self, debug=debug)
         if get_ipython() is not None:
-            frontend = frontends.JupyterFrontend(debug=debug)
+            frontend = frontends.JupyterFrontend()
         else:
-            frontend = frontends.ConsoleFrontend(debug=debug)
+            frontend = frontends.ConsoleFrontend()
         runner.set_frontend(frontend)
         frontend.set_runner(runner)
         runner.run()
@@ -462,34 +456,63 @@ class ParametrizedExercise:
 
 class ExerciseRunner:
 
-    def __init__(self, exercise, global_parameters={}):
+    def __init__(self, exercise, debug=False, global_parameters={}):
+        self.debug = debug
         self.observers = []
         self.pexercise = ParametrizedExercise(exercise, global_parameters)
+        self.sender = exercise.__class__
+        self.widget_id_mapping = {
+            widget.ID: widget for widget in self.pexercise.widgets
+        }
 
     # TODO: enforce order of steps
     def run(self):
-        self.frontend.render_preamble(self.pexercise.preamble)
-        self.frontend.render_problem(self.pexercise.template)
-        self.frontend.get_answers()
+        self.notify(ExerciseAttribute(self.sender, 'debug', self.debug))
+        self.notify(ExerciseAttribute(
+            self.sender, 'parameters', self.pexercise.parameters
+        ))
+        self.notify(RenderTemplate(
+            self.sender, 'preamble', self.pexercise.preamble
+        ))
+        for widget in self.pexercise.widgets:
+            self.notify(CreateWidget(
+                self.sender, widget.ID, widget.__class__.__name__
+            ))
+            widget.observe_attributes()
+            self.notify(ChangeWidgetAttribute(
+                self.sender, widget.ID, 'info', widget.info
+            ))
+        self.notify(RenderTemplate(
+            self.sender, 'problem', self.pexercise.template
+        ))
+        if self.debug:
+            self.publish_solutions()
+        self.notify(WaitingForSubmission(self.sender))
 
     def finish(self):
-        self.get_solutions()
-        self.notify(
-            self.pexercise.__class__, 'total_max_score',
+        if not self.debug:
+            self.publish_solutions()
+        self.notify(ExerciseAttribute(
+            self.sender, 'answers', self.pexercise.answers
+        ))
+        self.notify(ExerciseAttribute(
+            self.sender, 'max_total_score',
             process_total_score(self.pexercise.max_total_score)
-        )
-        self.notify(
-            self.pexercise.__class__, 'total_score',
+        ))
+        self.notify(ExerciseAttribute(
+            self.sender, 'total_score',
             process_total_score(self.pexercise.total_score)
-        )
+        ))
         self.pexercise.correct
         for widget in self.pexercise.model.widgets:
             widget.show_max_score = True
             widget.show_score = True
             widget.show_correct = True
-        self.frontend.render_feedback(self.pexercise.feedback)
+        self.notify(RenderTemplate(
+            self.sender, 'feedback', self.pexercise.feedback
+        ))
 
-    def get_solutions(self):
+    def publish_solutions(self):
         self.pexercise.solution
         for widget in self.pexercise.model.widgets:
             widget.show_solution = True
@@ -501,37 +524,18 @@ class ExerciseRunner:
         self.observers.append(observer)
         for widget in self.pexercise.widgets:
             widget.register_observer(observer)
-            widget.observe_attributes()
 
-    def notify(self, owner, name, value):
+    def notify(self, msg):
         for observer in self.observers:
-            observer(self, owner, name, value)
+            observer(msg)
 
-    @cache
-    def get_parameters(self):
-        return self.pexercise.parameters
-
-    def get_answers(self):
-        return self.pexercise.answers
-
-    @property
-    def widgets(self):
-        return self.pexercise.widgets
-
-    @property
-    def exercise_dir(self):
-        return self.pexercise.exercise.dir
-
-    def set_widget_value(self, widget, value):
-        widget.value = value
-
-    @property
-    def empty_widgets(self):
-        return [widget for widget in self.widgets if widget.value is None]
-
-    @property
-    def invalid_widgets(self):
-        return [widget for widget in self.widgets if widget.valid is False]
+    def observer(self, msg):
+        if isinstance(msg, ChangeWidgetAttribute):
+            if msg.attribute_name == 'value':
+                widget = self.widget_id_mapping[msg.widget_id]
+                widget.value = msg.attribute_value
+        elif isinstance(msg, Submit):
+            self.finish()
 
 
 class ExercisePool(collections.UserList):