+++ docs/changelog.rst
@@ -35,6 +35,7 @@ Changes
 * Rename :code:`score_types` to :code:`float_types`.
 * Create a :code:`nodes` package and outsource errors into a separate module to avoid circular imports.
 * Composed input fields can be invalid even if all children nodes are valid. Therefore all children nodes are now invalid if the composed input field is invalid.
+* Messages sent between frontends and runners are now encapsulated by the class :code:`Message`.
 
 Fixes
 -----